Add option pre_process_include_non_pdf_files.
Add option 'pre_process_include_non_pdf_files' which is false by default. For
processing the Fairfax ftp directory we only want to process pdf files in a
specific date range and ignore files that don't match the fairfax file format.
When processing other folders (such as for deduping and organization efforts)
we want to process all the files.

fairfax-pre-and-post-process-grouper.py:
- Add option pre_process_include_non_pdf_files. False by default (if it has not
  been set).
- More accurate display of pre-processed counts.
- Correct some formatting issues and clarify some output.