Attempt 5 times to get md5 sum with small delays.
There's an issue where a md5 check is failing on SunOS -- perhaps because the
mapped file share is not present enough (the file handle disappears). This is
showing up as an error 'digest: error reading file: I/O error' with a
CallProcessError. Subsequent attempts with the same file succeed, indicating
that the I/O is an intermittent issue, and not related to the program itself.
To work around this issue we attempt to get the md5 sum after a small dely.

fairfax-pre-and-post-process-grouper.py:
- get_md5_sum attempts up to 5 times to get the md5_sum (with a small delay
  each time).
- Reorder imports alphabetically.