Some os.walk debugging.
Put in some debugging to resolve a looping error when performing a os.walk
on some subdirectories.

It turned out that the error was due to a file with unicode characters that
was not being translated correctly through a cifs/samba mount, which meant
the file was linking to its parent directory, causing the os.walk to loop
endlessly. However, the debug messages and error handling on the os.walk
are left in place.